Separate EPGEvent class
